tests/crypto_aesctr: use _stream(), not _buf()
When we used crypto_aesctr_buf(), we were doing memory allocation and
calling insecure_memzero() for every single rep.  Over 20% of the time
was spent just in insecure_memzero()[1]!

[1] That's amortized over all the buffer sizes.  For a buffer size of
1024, it wasn't a big deal.  But when we were testing with a buffer size
of 1, we'd run insecure_memzero() on a 40-byte buffer for every single
byte that we encrypted in the test.

This commit moves _init() and _free() into the untimed perftest_init(),
instead of the timed perftest_func().