cpusupport: arm crc32 build
Definitions for our ARM support:
- we use "ARM" to refer to both ARM (32-bit) and AARCH64 (64-bit),
  because the platforms are sufficiently similar (from the perspective
  of C source code).
- the ARM C Language Extension (ACLE) is a set of headers for using
  their CPU intrinsics.

Compiler support for ARM is a bit spotty.  In general, we're looking for
compiler-defined macros, such as we see from

    clang -dM -E - < /dev/null | grep "__ARM" | sort

For CRC32 in particular, we want two macros:
    __ARM_ACLE: to indicate that we can #include <arm_acle.h> [1].
    __ARM_FEATURE_CRC32: to indicate that the compiler can understand
        functions such as __crc32b() and __crc32cd().

As of clang 11.0.0, __ARM_FEATURE_CRC32 is not enabled with
-march=armv8.1-a, or in fact with any -march option that does not
manually include a "+crc" [2].  CRC32 was optional in armv8-a (when
clang first added support for it), but it's mandatory in armv8.1-a and
up.

As of gcc 10.2, we are missing __ARM_ACLE, which is required before
including <arm_acle.h> [1].  I suspect this is because it is documented
as not fully supporting the Arm C Language Extension (ACLE) [3].
There's a related bug report [4].

That said, [3] does say that gcc fully supports the CRC32 intrinsics,
and empirically it seems to work if we add __ARM_ACLE ourselves.

[1] The ARM C language extensions docs say:
        The __ARM_ACLE macro should be tested before including the header:

        #ifdef __ARM_ACLE
        #include <arm_acle.h>
        #endif /* __ARM_ACLE */

    First paragraph in the "Header files" section of
    https://developer.arm.com/documentation/101028/0012/3--C-language-extensions?lang=en

[2] https://bugs.llvm.org/show_bug.cgi?id=33480

[3] https://gcc.gnu.org/onlinedocs/gcc-10.2.0/gcc/ARM-C-Language-Extensions-_0028ACLE_0029.html

[4] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=70974