align_ptr.h: add
These macros are an abstraction of code in crypto_aes_aesni.c.  They
give us a way to get a pointer to an array which is aligned to a desired
memory address alignment.  This is primarily intended for 16-byte
boundaries for SSE or ARM intrinsics.

There are various existing functions which deal with aligned memory, but
they all have problems:

- posix_memalign() is [ADV], which is an optional part of the standard.
- aligned_alloc() is C11.
- _mm_malloc() and _mm_free() are for x86 intrinsics only

If we ever decide to require posix_memalign or C11, the macros can be
adapted to use those methods.

Colin's clarification: The semantics required for uintptr_t are far
weaker -- the values don't have to be meaningful at all, beyond
returning the same pointer if you cast them back -- but every platform
I'm aware of provides this (I think even including CHERI, which breaks
all sorts of pointer-related assumptions).