sha256_sse2: use the BBAA rotate trick for s1
This optimizes
    s1(i) = (W[i-2] >>> 17) ^ (W[i-2] >>> 19) ^ (W[i-2] >> 10)
where >>> is a right-rotate.

In the context of MSG4, we can only calculate two such values at once:
s1(j-2) & s1(j-1).  We then add those to X4 and calculate s1(j) & s1(j+1).
As such, there are two unused 32-bit lanes.

As noted in [1, p. 10], we can duplicate the 32-bit values A and B to form:
    B, B, A, A
If we do a right-shift r() on both pairs of 64 bits, we have:
    x, r(B), x, r(a)
where x indicates garbage.  We can then shuffle r(B) and r(A) into the
desired lanes.

This gives us a speed boost on gcc and clang of between +1% to +5%.

[1] J. Guilford, K. Yap, V. Gopal.  "Fast SHA-256 Implementations on
    Intel(r) Architecture Processors", May 2012.
    https://www.intel.com/content/dam/www/public/us/en/documents/white-papers/sha-256-implementations-paper.pdf