sha256_sse2: MSG4 with rotating message schedule state
Optimize by reducing the number of memory loads needed.

This adds a new set of variables: Y[4], which are similar to the Xn
variables used in MSG4.  At the beginning of each main for loop (where
j=16, 32, 48, 64), they are defined as:
    Y[0] = W[j - 16] : W[j - 13]    (= X0)
    Y[1] = W[j - 12] : W[j - 9]     (= X1)
    Y[2] = W[j - 8] : W[j - 5]      (= X2)
    Y[3] = W[j - 4] : W[j - 1]      (= X3)

However, after the first MSG4 function call (during which we calculated
the next four values of W), we change the value of Y[0] to be:
    Y[0] = W[j + 0] : W[j + 3]      (= the previous X4)

We therefore rotate "the previous 16 values of W" through Y.

Note that we still need to store the final values in W, because the
round calculation does not use SSE and therefore requires the values to
be in main memory.