build: always expand variables before printing
s/-V/-v/g

The behaviour of `make -V FOO` depends on .MAKE.EXPAND_VARIABLES,
using -dV, and probably the contents of some stuff in /usr/share/mk/.
In other words, it's a mess.

By contrast, `make -v FOO` is documented as always expanding the
variable.  That's what we want.