Actions: adjust clang warnings for pthread_mutex_lock()
Clang's thread safety analysis warns that it cannot handle
"conditionally held locks".  They give the example of:
    bool b = needsToLock();
    if (b) mu.Lock();
    ...  // Warning!  Mutex 'mu' is not held on every path through here.
    if (b) mu.Unlock();

This is rather unfortunate, because pthread_mutex_lock() is documented
as being able to fail and set errno [2].  Well-written code will
therefore not assume that the lock is successful, and will instead check
the returned value and jump to error handling if necessary.  In other
words, well-written code will include "conditionally held locks".

[1] https://clang.llvm.org/docs/ThreadSafetyAnalysis.html#conditional-locks
[2] https://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_mutex_trylock.html