crypto_aes_arm: add software-only (with openssl)
To keep the history of crypto_aes_arm.c as clear as possible, these functions
are exact copies of the generic code (which uses OpenSSL), except for removing
the "what type of intrinsics should we use?" code at the top.  The OpenSSL
calls will be replaced with ARM-AES intrinsics in the next PR in this series;
we're only using these temporarily, so that we can reach the "compiles and
passes the tests" stage with a minimum of fuss.

(We used this strategy in 192d05beba9502601496102ee89258afc534387d and
c06b8e3589b988377608b852b26d95d6c569aef5 as well.)