network_ssl_compat: add LibreSSL support for 2.7.0+
LibreSSL earlier than 2.7.0 (which was 2018-03-21) is not supported.
The problem is hostname validation.  OpenSSL provides an example on
    https://wiki.openssl.org/index.php/Hostname_validation
which does not work on LibreSSL, which is missing
    SSL_get0_param()
    X509_VERIFY_PARAM_set1_host()
    X509_VERIFY_PARAM_set_hostflags

It may be possible to replicate this with other functions.  However,
LibreSSL appears to move fairly quickly:
    Support Schedule

    LibreSSL transitions to a new stable release branch every 6 months
    in coordination with the OpenBSD development schedule. LibreSSL
    stable branches are updated for 1 year after their corresponding
    OpenBSD branch is tagged for release.
    https://www.libressl.org/releases.html

LibreSSL 2.7.2 corresponded to OpenBSD 6.3 (released 2018-04-02; EOL
2019-05-03); the last upgrade to 2.7.x was 2.7.5 on 2018-12-15.
LibreSSL 2.6.5 was the last upgrade to the 2.6.x line, and was on
2018-05-13.  As such, I'm not fussed about not supporting 2.6.5 or
earlier, unless/until somebody has a pressing need for that.