tests/buildsingles: use IDIRS
Previously, we would pass the exact relative filename of each header, so
after pre-processing, main.c would contain a line in the form:
    #include "../../aws/aws_sign.h"
This changes that to be:
    #include "aws_sign.h"
and passing -I../../aws to the command-line, which is how we normally
handle includes.

As an added bonus, this allows our header files to include other
libcperciva header files.  Previously, this test could only handle
included system header files.