elasticarray: rewrite prefix##_iter(), remove _iter()
The previous prefix##_iter() was invoking undefined behaviour.  Given
two functions:
    void fp_one(rectype *)
    void fp_two(void *)
prefix##_iter() was accepting a function pointer in the form of fp_one,
then casting it to the form of fp_two and passing it to
elasticarray_iter(), which used it in that form.  This violated:
    A pointer is used to call a function whose type is not compatible
    with the pointed-to type (6.3.2.3).
    - C99 Annex J.2 Undefined behaviour

It would have been ok if elasticarray_iter() had casted fp_two back to
fp_one:
    A pointer to a function of one type may be converted to a pointer to
    a function of another type and back again; the result shall compare
    equal to the original pointer. If a converted pointer is used to
    call a function whose type is not compatible with the pointed-to
    type, the behavior is undefined.
    - C99 6.3.2.3 Pointers, paragraph 8
but in order to do that, elasticarray_iter() would need to know the
rectype, and couldn't take a (void *).

Another solution would be to have elasticarray_iter() invoke a callback
prefix##_iter_callback, which casts from (void *) back to (rectype *)
before calling the user-provided callback.

The rewritten prefix##_iter() now acts on (rectype *) directly.  As
such, there's no need for elasticarray_iter(), so it's been removed.