tests: shorten some of the names in the public API
A later commit will switch to using "local" variables (indicated by
prepending the variable name with the function name), but those would be
extremely long with the previous function names.  Shortening them will
result in variables which are merely very long.

Changes to the overall API:
    s/setup_check_variables/setup_check/g

Changes to the valgrind API:
    s/valgrind_setup_cmd/valgrind_setup/g
    s/valgrind_check_basenames/valgrind_check/g