tests/mpool: allocate 128 bytes instead of 4
This is needed for FreeBSD 14.0 with clang; their malloc is allocating a
slab of pointers at a time, which benefits small allocations.

Our mpool test is not intended as a serious benchmark of memory
allocation speeds; rather, it's a quick check that nothing unusual is
happening with the platform's kernel and libc.  For a serious
comparison, we would need to check the actual scenarios: for example,
are we allocating 4096 records of 16 bytes, or 16 records of 52 bytes?